+++ src/main/java/com/oneDev/healthcarebooking/entity/HospitalDoctorFee.java
@@ -8,6 +8,7 @@ import lombok.NoArgsConstructor;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -29,7 +30,7 @@ public class HospitalDoctorFee {
     private Long doctorSpecializationId;
 
     @Column(nullable = false)
-    private Double fee;
+    private BigDecimal fee;
 
     @CreationTimestamp
     @Column(name = "created_at", nullable = false, updatable = false)
